Extracted pricing slider and card to resuable component that is to be reused in billing flow
